mysite/requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-filter==2.4.0
django-modelcluster==5.1
django-recaptcha==2.0.6
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
isort==5.6.4
jdcal==1.4.1
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.5
Pillow==7.2.0
pylint==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
Unidecode==1.1.1
urllib3==1.25.11
wagtail==2.10.2
wagtail-django-recaptcha==1.0
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.4
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
